refactor: divide csv loading to other package
