+++ src/lib.rs
@@ -6,3 +6,4 @@ pub mod linear_algebra;
 pub mod lu_decomposition;
 pub mod newton;
 pub mod quasi_newton;
+pub mod utils;
